Make #169 backwards-compatible
We probably never meant for `guid` to be used, so this is probably a
bug, but its been a part of the public interface for so long that I
don't feel right changing it with just a bug-fix release.

Instead, I'm going to treat this as a feature and this commit makes it
backwards-compatible. Once we go 1.0, I'll drop guid altogether.