Fixes https://github.com/feedjira/feedjira/issues/212
When the Atom parser is deciding which of the <link> elements present holds the URL for the feed, look for a <link> with rel="self". When deciding which is the website URL, if no <link> has type="text/html", return the href value for the last <link> in the XML which is not the feed <link>, if possible.