Use vcr to stub network requests (#325)
there are three tests in feed_spec that require network to pass.

PR #297 fixes the immediate issue by making the test general enough to
pass when the response from the server changes somewhat.

PR #300 fixes the network issue, but uses specific stubs. I would say
this method is no better or worse than this one, however it does require
knowing that you are stubbing the request and manually keeping responses
up to date.

This method acts the most like actually hitting the network and will
work on an airplane.