Fixes https://github.com/feedjira/feedjira/issues/327
Changed AtomFeedBurner parser so that it correctly parses feedburner atom feeds in the new style some feeds seem to be using. In this alternate style:

- the link to the feed webpage may be a plain <link> element without the type="text/html" attribute we usually expected
- the link to the feed itself may be a <atom10:link> element instead of the <link> element we usually expected (with type="application/atom+xml")

The way the change is implemented, old-style values will take precedence if present in the feed. Only if the new style is used will the parser use the new locations for the url and feed_url values.