Fix Feedburner Itunes feed parsing
When an Itunes feed is published on Feedburner, it parses it as a
Feedburner feed, which doesn't capture the enclosure url for the podcast
download. This change moves Itunes feed parsing above Feedburner, so
that for these cases, the feed will instead be parsed as an Itunes feed
rather than a Feedburner feed. This will ensure that the enclosure url
will be available.

Fixes [#464][464] and probably fixes [#428][428].

[464]: https://github.com/feedjira/feedjira/issues/464
[428]: https://github.com/feedjira/feedjira/issues/428