Increasing slice limit regarding able_to_parse?
I encountered a feed that had a very long base64 encoded string from https://www.rss.style at the beginning of is RSS XML file.

The previous limit of 2000 characters was sufficient and in turn able_to_parse? always returned "false".

This commit doubles that limit to 4000 characters, which was sufficient to parse the feed (which otherwise was a valid RSS feed without problems).